Ecommerce Gym Equipment Website - 4th Year/App/backend/requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-cors-headers